Add 'target_os = ["win"]' to the webrtc .gclient config.
Also check for the existence of .gclient, rather than src, when deciding
whether to initialize. I'm doing this because I suspect the addition of
target_os will not take effect for anyone who already has a sync of
webrtc, and it's better if they can reinitialize just be deleting
.gclient, rather than deleting the whole src directory.

The funny-looking unindent is necessary because the "target_os" has to
be at the beginning of a line; otherwise gclient throws a syntax error.